Refactoring publisher user service
