+++ MonolitApp/agent-app/src/main/java/agent/app/service/impl/AdServiceImpl.java
@@ -295,58 +295,4 @@ public class AdServiceImpl implements AdService {
         return ads;
     }
 
-    @Override
-    public AdStatisticsDTO findBestAverageGrade(String email) {
-        Ad adT = null;
-        double averageGrade = 0.0;
-        double max = 0.0;
-        System.out.println("Average method");
-        for (Ad ad : adRepository.findAll()) {
-            if (ad.getRatingCnt() == 0) {
-                continue;
-            } else {
-                averageGrade = ad.getRatingNum() / ad.getRatingCnt();
-                if (averageGrade > max) {
-                    System.out.println("Average: " + averageGrade);
-                    max = averageGrade;
-                    adT = ad;
-                }
-            }
-        }
-        AdStatisticsDTO adPage = AdConverter.toCreateAdStatisticsDTOFromAd(adT);
-        return adPage;
-    }
-
-    @Override
-    public Ad findAdWithGrade(Double max_grade) {
-        System.out.println("Metodaaa averg... ");
-        for (Ad ad : adRepository.findAll()) {
-            if ((ad.getRatingNum() / ad.getRatingCnt()) == max_grade) {
-                System.out.println(max_grade);
-                return ad;
-            }
-        }
-        return null;
-    }
-
-    @Override
-    public AdStatisticsDTO findMaxMileage(String email) {
-        Ad adT = null;
-        float max = 0;
-        System.out.println("Average method za kilometrazu");
-        for (Ad ad : adRepository.findAll()) {
-            if (ad.getCar().getMileage() > max) {
-                System.out.println("Max km: " + ad.getCar().getMileage());
-                max = ad.getCar().getMileage();
-                adT = ad;
-            }
-
-
-        }
-
-        AdStatisticsDTO adPage = AdConverter.toCreateAdStatisticsDTOFromAd(adT);
-        return adPage;
-    }
-
-
 }
