+++ MonolitApp/agent-app/src/main/java/agent/app/service/intf/PublisherUserService.java
@@ -1,5 +1,6 @@
 package agent.app.service.intf;
 
+import agent.app.dto.ad.AdStatisticsDTO;
 import agent.app.dto.car.CarCalendarTermCreateDTO;
 import agent.app.model.CarCalendarTerm;
 import agent.app.model.PriceList;
@@ -20,4 +21,6 @@ public interface PublisherUserService {
     Boolean existsByEmail(String email);
     PublisherUser findByEmail(String email);
     List<PriceList> findPriceListsFromPublishUser(String publishUserUsername);
+    AdStatisticsDTO findBestAverageGrade(String email);
+    AdStatisticsDTO findMaxMileage(String email);
 }
